==> 6-APP.rmd <==
---
title: "Dashboard SDG"
author: "Victoria Costas"
date: "`r Sys.Date()`"
output:
  html_document:
    toc: yes
---

```{r global_options, include=T, echo = F}
knitr::opts_chunk$set(echo = T, warning=FALSE, message=FALSE)
```

```{=html}
<style>
body {
text-align: justify}
</style>
```

```{r}
library(tidyverse)
library(shiny)
library(shinydashboard)
library(shinydashboardPlus)
library(base64enc)
library(plotly)
library(factoextra)
```

## Data needed for the Dashboard

```{r}
spain_trends <- read_csv("./data/spain_trends.csv")
germany_trends <- read_csv("./data/germany_trends.csv")

#remove the first row for clarity
spain_trends <- spain_trends[, -1]
germany_trends <- germany_trends[, -1]
```

```{r}
trends <- rbind(spain_trends, germany_trends)

# Scale the "value" column in the "trends" dataset
trends_scaled <- trends 
trends_scaled$value <- scale(trends$value)
```

```{r}
#compute the average for each goal each year
average_by_goal_year <- trends_scaled %>%
  group_by(indicator, year, geo_time) %>%
  summarise(Average_Value = mean(value, na.rm = TRUE))
```

```{r}
#data for the year 2010 onwards
data <- average_by_goal_year %>% filter(year >= 2010)
```

```{r}
overall_stats_es <- data %>% filter(geo_time== "ES") %>% 
  group_by(indicator, year) %>%
  summarise(Average_Value = Average_Value, Max_Value = max(Average_Value, na.rm = T), Min_Value = min(Average_Value, na.rm = T))

overall_stats_de <- data %>% filter(geo_time == "DE") %>% 
  group_by(indicator, year) %>%
  summarise(Average_Value = Average_Value, Max_Value = max(Average_Value, na.rm = T), Min_Value = min(Average_Value, na.rm = T))

# Combine overall_stats_es and overall_stats_de into a single data frame
overall_stats_combined <- rbind(overall_stats_es, overall_stats_de)
```

```{r}
data_mf <- read_csv("data/data_mf.csv")
data_mf <- data_mf[,-1]
#Entire time series
X <- data_mf %>% select(-c(year, geo_time))
set.seed(123)
k = 7
fit <- kmeans(scale(X), centers = k, nstart = 10000)
groups <- fit$cluster
cluster_labels <- fit$cluster
names <- data_mf$geo_time
centers <- fit$centers

#From 2010 to 2015
data_old <- subset(data_mf, year %in% c(2010, 2011, 2012, 2013, 2014, 2015))
X_old <- data_old %>% select(-c(year, geo_time))

k = 7
fit_old <- kmeans(scale(X_old), centers = k, nstart = 10000)
groups_old <- fit_old$cluster
cluster_labels_old <- fit_old$cluster
names_old <- data_old$geo_time
centers_old <- fit_old$centers


#From 2016 to 2021
data_new <- subset(data_mf, year %in% c(2016, 2017, 2018, 2019, 2020, 2021))
X_new <- data_new %>% select(-c(year, geo_time))

k = 6
fit_new <- kmeans(scale(X_new), centers = k, nstart = 10000)
groups_new <- fit_new$cluster
cluster_labels_new <- fit_new$cluster
names_new <- data_new$geo_time
centers_new <- fit_new$centers


```

## Functions needed for the Dashboard

```{r}
cluster_plot <- function(year){
  set.seed(123)
  if (year == "2010-2021"){
    
      p <- fviz_cluster(fit, data = X, geom = c("point"), ellipse.type = 'norm', pointsize=1)+
      theme_minimal()+ geom_text(label=names,hjust=0, vjust=0,size=2,check_overlap = T) + scale_fill_brewer(palette="Paired")
      
      clusplot <- ggplotly(p, tooltip = "text")
      
  } else if (year == "2010-2015 (Pre-SDG period)"){
    
      p_old <- fviz_cluster(fit_old, data = X_old, geom = c("point"), ellipse.type = 'norm', pointsize=1)+
      theme_minimal()+ geom_text(label=names_old,hjust=0, vjust=0,size=2,check_overlap = T) + scale_fill_brewer(palette="Paired")
      
      clusplot <- ggplotly(p_old, tooltip = "text")
      
  } else if (year == "2016-2021 (SDG period)") {
    
      p_new <- fviz_cluster(fit_new, data = X_new, geom = c("point"), ellipse.type = 'norm', pointsize=1)+
      theme_minimal()+ geom_text(label=names_new,hjust=0, vjust=0,size=2,check_overlap = T) + scale_fill_brewer(palette="Paired")
      
      clusplot <- ggplotly(p_new, tooltip = "text")
  }
 
   return(clusplot)
}
```

```{r}
bar_image <- function(year, is_first){
  if (year == "2010-2021"){
    if(is_first){
      image_file <- "www/6_all.png"
    } else{
      image_file <- "www/7_all.png"
    }
  } else if (year == "2010-2015 (Pre-SDG period)"){
    if(is_first){
      image_file <- "www/5_old_es.png"
    } else{
      image_file <- "www/3_old_de.png"
    }
  } else if (year == "2016-2021 (SDG period)") {
    if(is_first){
      image_file <- "www/5_new_both.png"
    } else{
      return("")
    }
  }
   return(image_file)
}


bar_text <- function(year){
  if (year == "2016-2021 (SDG period)") {
    text <- ""
  } else {
    text <- "Goal Performance (Germany)"
  }
   return(text)
}
```


```{r}
plot_trends <- function(indicator_number){
  renderPlot({
    data %>% filter(indicator == paste0("ind", indicator_number)) %>% 
  ggplot(aes(x = year, y = Average_Value, color = geo_time, group = geo_time)) +
  geom_line(size = 1.5) +
  labs(title = paste0("Progress of Goal ", indicator_number) , x = "Year", y = "Average Value") +
  theme_minimal() +
  theme(plot.title = element_text(size = 16, face = "bold"),
        axis.title = element_text(size = 12),
        axis.text.x = element_text(angle = 45, hjust = 1, size = 10),
        axis.text.y = element_text(size = 10)) +
  scale_color_manual(values = c("ES" = "blue", "DE" = "red"),
                     labels = c("ES" = "Spain", "DE" = "Germany"),
                     name = "Country")
  })
}

```

```{r}
 output_goals <- function(i) {
   image_file <- paste0("www/SDG", i, ".png")
   image_data <- readBin(image_file, "raw", file.info(image_file)$size)
   image <- paste0("data:image/jpeg;base64,", base64encode(image_data))
   return(image)
 }

file <- "www/SDG_all.png"
image_data_principal <- readBin(file, "raw", file.info(file)$size)
image_principal <- paste0("data:image/jpeg;base64,", base64encode(image_data_principal)) 
```

```{r}
top_countries_by_goal <- function(){
    result <- list()
    
    for (ind in 1:17) {
      goal <- paste0("ind", ind)
      
      #Filter data for 2010
      filtered_year_2010 <- subset(data_mf, year == "2010")
      top_10_2010 <- subset(filtered_year_2010, select = c("geo_time", goal))
      top_10_2010 <- top_10_2010[order(-top_10_2010[[goal]]), ]
      top_10_2010 <- head(top_10_2010, 10)
      
      #Filter data for 2020
      filtered_year_2020 <- subset(data_mf, year == "2020")
      top_10_2020 <- subset(filtered_year_2020, select = c("geo_time", goal))
      top_10_2020 <- top_10_2020[order(-top_10_2020[[goal]]), ]
      top_10_2020 <- head(top_10_2020, 10)
      
      goal_df <- data.frame("2010" = top_10_2010$geo_time, "2020" = top_10_2020$geo_time, check.names = FALSE)
      result[[goal]] <- goal_df
    }
    
    return(result)
  }
```


## Dashboard components

```{r}
header <- dashboardHeader(title = "SDG in Spain and Germany",
                          titleWidth = 300)
```


```{r}
sidebar <- dashboardSidebar(
  width = 300,
  sidebarMenu(
    menuItem("Overview", tabName = "Overview", icon = icon("dashboard")),
    menuItem("Goals", icon = icon("bullseye"),
             menuSubItem("Goal 1", tabName = "Goal1"),
             menuSubItem("Goal 2", tabName = "Goal2"),
             menuSubItem("Goal 3", tabName = "Goal3"),
             menuSubItem("Goal 4", tabName = "Goal4"),
             menuSubItem("Goal 5", tabName = "Goal5"),
             menuSubItem("Goal 6", tabName = "Goal6"),
             menuSubItem("Goal 7", tabName = "Goal7"),
             menuSubItem("Goal 8", tabName = "Goal8"),
             menuSubItem("Goal 9", tabName = "Goal9"),
             menuSubItem("Goal 10", tabName = "Goal10"),
             menuSubItem("Goal 11", tabName = "Goal11"),
             menuSubItem("Goal 12", tabName = "Goal12"),
             menuSubItem("Goal 13", tabName = "Goal13"),
             menuSubItem("Goal 14", tabName = "Goal14"),
             menuSubItem("Goal 15", tabName = "Goal15"),
             menuSubItem("Goal 16", tabName = "Goal16"),
             menuSubItem("Goal 17", tabName = "Goal17")),
    menuItem("Notes", tabName = "Notes", icon = icon("sticky-note"),
             badgeLabel = "new", badgeColor = "yellow")
  )
)
```


```{r}
body <- dashboardBody(
  tags$style(HTML("
      .menu-subitem .title {
        display: flex;
        align-items: center;
      }
      .menu-subitem .title img {
        margin-right: 5px;
        margin-left: 5px;
        height: 10%;
      }
    ")),
  tabItems(
    tabItem(tabName = "Overview", 
            # h2("Overview"),
             fluidPage(
              titlePanel(
                div(
                    img(src = image_principal, height = "100px"),
                    "Overview",
                    class = "title"
                  )
                )
              ),
            fluidRow(
                column(
                  width = 4,
                  selectInput(
                    inputId = "yearFilter",
                    label = "Select Year",
                    choices = c("2010-2015 (Pre-SDG period)", "2016-2021 (SDG period)", "2010-2021"),
                    selected = "2010-2015 (Pre-SDG period)"
                  )
                )
            ),
            fluidRow(
              column(
                  width = 8,
                  plotlyOutput("clusplot")
              )
            ),
            fluidRow(
              column(
                  width = 8,
                  div(
                    HTML("<br><br>"),
                    h3("Goal Performance"),
                    plotOutput("barplot1")
                  )
              )
            ),
            fluidRow(
              column(
                  width = 8,
                  div(
                    h3(textOutput("barplot2text")),
                    plotOutput("barplot2")
                  )
              )
            ),
    ),
    tabItem(tabName = "widgets",
      h2("Widgets tab content")
    ),
    tabItem(tabName = "Notes",
      h2("Notes on the Dashboard"),
      fluidRow(
        column(
        width = 4,
        box(
          title = div("Notices", class = "custom-box-header-pos"),
          width = 12,
          solidHeader = TRUE,
          p("All data shown here are taken from Eurostat's Sustainable Development Goals database, but have been further processed. The results shown here may vary from data from organisations such as the UN or the EU, because the UN does not specify how they measure goals. This dashboard is only intended to show the performance of Germany and Spain and their context.")
          )
        )
      )
    ),
    tabItem(tabName = "Goal1",
            fluidPage(
              titlePanel(
                div(
                    img(src = output_goals('1'), height = "100px"),
                    "Goal 1. No Poverty",
                    class = "title"
                  )
                )
              ),
                fluidRow(
                  box(width = 12, title = "Progress Plot and Top-10 Countries in 2010 and 2020", 
                    splitLayout(
                      cellWidths = c("75%", "25%"),
                      plotOutput("plot1"),
                      tableOutput("table_goal1")
                    )
                )
            )
      ),
    tabItem(tabName = "Goal2",
            fluidPage(
              titlePanel(
                div(
                    img(src = output_goals('2'), height = "100px"),
                    "Goal 2. Zero Hunger",
                    class = "title"
                  )
                )
              ),
                fluidRow(
                  box(width = 12, title = "Progress Plot and Top-10 Countries in 2010 and 2020", 
                    splitLayout(
                      cellWidths = c("75%", "25%"),
                      plotOutput("plot2"),
                      tableOutput("table_goal2")
                    )
                )
            )
      ),
    tabItem(tabName = "Goal3",
            fluidPage(
              titlePanel(
                div(
                    img(src = output_goals('3'), height = "100px"),
                    "Goal 3. Good Heath and Well-being",
                    class = "title"
                  )
                )
              ),
                fluidRow(
                  box(width = 12, title = "Progress Plot and Top-10 Countries in 2010 and 2020", 
                    splitLayout(
                      cellWidths = c("75%", "25%"),
                      plotOutput("plot3"),
                      tableOutput("table_goal3")
                    )
                )
            )
      ),
    tabItem(tabName = "Goal4", 
            fluidPage(
              titlePanel(
                div(
                    img(src = output_goals('4'), height = "100px"),
                    "Goal 4. Quality Education",
                    class = "title"
                  )
                )
              ),
                fluidRow(
                  box(width = 12, title = "Progress Plot and Top-10 Countries in 2010 and 2020", 
                    splitLayout(
                      cellWidths = c("75%", "25%"),
                      plotOutput("plot4"),
                      tableOutput("table_goal4")
                    )
                )
            )
      ),
    tabItem(tabName = "Goal5",
            fluidPage(
              titlePanel(
                div(
                    img(src = output_goals('5'), height = "100px"),
                    "Goal 5. Gender Equality",
                    class = "title"
                  )
                )
              ),
                fluidRow(
                  box(width = 12, title = "Progress Plot and Top-10 Countries in 2010 and 2020", 
                    splitLayout(
                      cellWidths = c("75%", "25%"),
                      plotOutput("plot5"),
                      tableOutput("table_goal5")
                    )
                )
            )
      ),
    tabItem(tabName = "Goal6",
            fluidPage(
              titlePanel(
                div(
                    img(src = output_goals('6'), height = "100px"),
                    "Goal 6. Clean Water and Sanitation",
                    class = "title"
                  )
                )
              ),
                fluidRow(
                  box(width = 12, title = "Progress Plot and Top-10 Countries in 2010 and 2020", 
                    splitLayout(
                      cellWidths = c("75%", "25%"),
                      plotOutput("plot6"),
                      tableOutput("table_goal6")
                    )
                )
            )
      ),
    tabItem(tabName = "Goal7",
            fluidPage(
              titlePanel(
                div(
                    img(src = output_goals('7'), height = "100px"),
                    "Goal 7. Affordable and Clean Energy",
                    class = "title"
                  )
                )
              ),
                fluidRow(
                  box(width = 12, title = "Progress Plot and Top-10 Countries in 2010 and 2020", 
                    splitLayout(
                      cellWidths = c("75%", "25%"),
                      plotOutput("plot7"),
                      tableOutput("table_goal7")
                    )
                )
            )
      ),
    tabItem(tabName = "Goal8", 
            fluidPage(
              titlePanel(
                div(
                    img(src = output_goals('8'), height = "100px"),
                    "Goal 8. Decent Work and Economic Growth",
                    class = "title"
                  )
                )
              ),
                fluidRow(
                  box(width = 12, title = "Progress Plot and Top-10 Countries in 2010 and 2020", 
                    splitLayout(
                      cellWidths = c("75%", "25%"),
                      plotOutput("plot8"),
                      tableOutput("table_goal8")
                    )
                )
            )
      ),
    tabItem(tabName = "Goal9", 
            fluidPage(
              titlePanel(
                div(
                    img(src = output_goals('9'), height = "100px"),
                    "Goal 9. Industry, Innovation and Infrastructure",
                    class = "title"
                  )
                )
              ),
                fluidRow(
                  box(width = 12, title = "Progress Plot and Top-10 Countries in 2010 and 2020", 
                    splitLayout(
                      cellWidths = c("75%", "25%"),
                      plotOutput("plot9"),
                      tableOutput("table_goal9")
                    )
                )
            )
      ),
    tabItem(tabName = "Goal10",
            fluidPage(
              titlePanel(
                div(
                    img(src = output_goals('10'), height = "100px"),
                    "Goal 10. Reduced Inequalities",
                    class = "title"
                  )
                )
              ),
                fluidRow(
                  box(width = 12, title = "Progress Plot and Top-10 Countries in 2010 and 2020", 
                    splitLayout(
                      cellWidths = c("75%", "25%"),
                      plotOutput("plot10"),
                      tableOutput("table_goal10")
                    )
                )
            )
      ),
    tabItem(tabName = "Goal11",
            fluidPage(
              titlePanel(
                div(
                    img(src = output_goals('11'), height = "100px"),
                    "Goal 11. Sustainable Cities and Communities",
                    class = "title"
                  )
                )
              ),
                fluidRow(
                  box(width = 12, title = "Progress Plot and Top-10 Countries in 2010 and 2020", 
                    splitLayout(
                      cellWidths = c("75%", "25%"),
                      plotOutput("plot11"),
                      tableOutput("table_goal11")
                    )
                )
            )
      ),
    tabItem(tabName = "Goal12",
            fluidPage(
              titlePanel(
                div(
                    img(src = output_goals('12'), height = "100px"),
                    "Goal 12. Responsible Consumption and Production",
                    class = "title"
                  )
                )
              ),
                fluidRow(
                  box(width = 12, title = "Progress Plot and Top-10 Countries in 2010 and 2020", 
                    splitLayout(
                      cellWidths = c("75%", "25%"),
                      plotOutput("plot12"),
                      tableOutput("table_goal12")
                    )
                )
            )
      ),
    tabItem(tabName = "Goal13", 
            fluidPage(
              titlePanel(
                div(
                    img(src = output_goals('13'), height = "100px"),
                    "Goal 13. Climate Action",
                    class = "title"
                  )
                )
              ),
                fluidRow(
                  box(width = 12, title = "Progress Plot and Top-10 Countries in 2010 and 2020", 
                    splitLayout(
                      cellWidths = c("75%", "25%"),
                      plotOutput("plot13"),
                      tableOutput("table_goal13")
                    )
                )
            )
      ),
    tabItem(tabName = "Goal14", 
            fluidPage(
              titlePanel(
                div(
                    img(src = output_goals('14'), height = "100px"),
                    "Goal 14. Life Below Water",
                    class = "title"
                  )
                )
              ),
                fluidRow(
                  box(width = 12, title = "Progress Plot and Top-10 Countries in 2010 and 2020", 
                    splitLayout(
                      cellWidths = c("75%", "25%"),
                      plotOutput("plot14"),
                      tableOutput("table_goal14")
                    )
                )
            )
      ),
    tabItem(tabName = "Goal15", 
            fluidPage(
              titlePanel(
                div(
                    img(src = output_goals('15'), height = "100px"),
                    "Goal 15. Life on Land",
                    class = "title"
                  )
                )
              ),
                fluidRow(
                  box(width = 12, title = "Progress Plot and Top-10 Countries in 2010 and 2020", 
                    splitLayout(
                      cellWidths = c("75%", "25%"),
                      plotOutput("plot15"),
                      tableOutput("table_goal15")
                    )
                )
            )
      ),
    tabItem(tabName = "Goal16",
            fluidPage(
              titlePanel(
                div(
                    img(src = output_goals('16'), height = "100px"),
                    "Goal 16. Peace, Justice and Strong Institutions",
                    class = "title"
                  )
                )
              ),
                fluidRow(
                  box(width = 12, title = "Progress Plot and Top-10 Countries in 2010 and 2020", 
                    splitLayout(
                      cellWidths = c("75%", "25%"),
                      plotOutput("plot16"),
                      tableOutput("table_goal16")
                    )
                )
            )
      ),
    tabItem(tabName = "Goal17", 
            fluidPage(
              titlePanel(
                div(
                    img(src = output_goals('17'), height = "100px"),
                    "Goal 17. Partnerships for the Goals",
                    class = "title"
                  )
                )
              ),
                fluidRow(
                  box(width = 12, title = "Progress Plot and Top-10 Countries in 2010 and 2020", 
                    splitLayout(
                      cellWidths = c("75%", "25%"),
                      plotOutput("plot17"),
                      tableOutput("table_goal17")
                    )
                )
            )
      )
  )
)


```

## Server and UI

```{r}
#Put them together into a dashboardPage
ui <- dashboardPage(
  header,
  sidebar,
  body
)
```

```{r}
server <- function(input, output) {
  #Render plots for Goals
  output$plot1 <- plot_trends('1')
  output$plot2 <- plot_trends('2')
  output$plot3 <- plot_trends('3')
  output$plot4 <- plot_trends('4')
  output$plot5 <- plot_trends('5')
  output$plot6 <- plot_trends('6')
  output$plot7 <- plot_trends('7')
  output$plot8 <- plot_trends('8')
  output$plot9 <- plot_trends('9')
  output$plot10 <- plot_trends('10')
  output$plot11 <- plot_trends('11')
  output$plot12 <- plot_trends('12')
  output$plot13 <- plot_trends('13')
  output$plot14 <- plot_trends('14')
  output$plot15 <- plot_trends('15')
  output$plot16 <- plot_trends('16')
  output$plot17 <- plot_trends('17')
  
  #Define the observer for the country and year filters
  observeEvent(
    {
      input$yearFilter
    },
    {
      #Update the plots based on the selected options
      output$clusplot <- renderPlotly(cluster_plot(input$yearFilter))
      
      output$barplot1 <- renderImage({
            filename <- bar_image(input$yearFilter, is_first = TRUE)
            #Return a list containing the file name
            list(src = filename,
                 width = "75%",
                 height = "auto")
          }, deleteFile=FALSE)
      output$barplot2 <- renderImage({
            filename <- bar_image(input$yearFilter, is_first = FALSE)
            #Return a list containing the filename
            list(src = filename,
                 width = "75%",
                 height = "auto")
          }, deleteFile=FALSE)
      output$barplot2text <- renderText(bar_text(input$yearFilter))
    }
  )
  
  top_countries_by_goal_results <- top_countries_by_goal()
  output$table_goal1 <- renderTable(top_countries_by_goal_results[[1]], bordered = T)
  output$table_goal2 <- renderTable(top_countries_by_goal_results[[2]], bordered = T)
  output$table_goal3 <- renderTable(top_countries_by_goal_results[[3]], bordered = T)
  output$table_goal4 <- renderTable(top_countries_by_goal_results[[4]], bordered = T)
  output$table_goal5 <- renderTable(top_countries_by_goal_results[[5]], bordered = T)
  output$table_goal6 <- renderTable(top_countries_by_goal_results[[6]], bordered = T)
  output$table_goal7 <- renderTable(top_countries_by_goal_results[[7]], bordered = T)
  output$table_goal8 <- renderTable(top_countries_by_goal_results[[8]], bordered = T)
  output$table_goal9 <- renderTable(top_countries_by_goal_results[[9]], bordered = T)
  output$table_goal10 <- renderTable(top_countries_by_goal_results[[10]], bordered = T)
  output$table_goal11 <- renderTable(top_countries_by_goal_results[[11]], bordered = T)
  output$table_goal12 <- renderTable(top_countries_by_goal_results[[12]], bordered = T)
  output$table_goal13 <- renderTable(top_countries_by_goal_results[[13]], bordered = T)
  output$table_goal14 <- renderTable(top_countries_by_goal_results[[14]], bordered = T)
  output$table_goal15 <- renderTable(top_countries_by_goal_results[[15]], bordered = T)
  output$table_goal16 <- renderTable(top_countries_by_goal_results[[16]], bordered = T)
  output$table_goal17 <- renderTable(top_countries_by_goal_results[[17]], bordered = T)
}

```

# Run the Dashboard APP
 
```{r}
shinyApp(ui, server)
```
